+++ app/src/main/java/de/passwordvault/model/autofill/ParsedStructure.java
@@ -1,0 +1,11 @@
+package de.passwordvault.model.autofill;
+
+import android.view.autofill.AutofillId;
+
+public class ParsedStructure {
+
+    public AutofillId usernameId;
+
+    public AutofillId passwordId;
+
+}