+++ app/src/main/java/de/passwordvault/model/autofill/UserData.java
@@ -1,0 +1,9 @@
+package de.passwordvault.model.autofill;
+
+public class UserData {
+
+    public String username;
+
+    public String password;
+
+}